adding swipe action to rotate the camera around something
