+++ Assets/Prefab/PlayerPrefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61743cabc1eaad848bd3cccbee05d55d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
