+++ Assets/Prefab/PlayerPrefabs/CameraRotationCenter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 19db3ea36af38354d8136bbddf7b51e6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
