+++ Assets/Prefab/PlayerPrefabs/Main Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 686ef2fe421bb1d4cab89eb34f3226fa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
