+++ Assets/Scripts/PlayerScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e925abcde38ccd40873b554987feef9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
