+++ Assets/Scripts/PlayerScript/HoldCamera.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 98dccff273e281c41bd71b7ac3ae8d7f
