+++ Packages/manifest.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "com.unity.adaptiveperformance.google.android": "5.1.2",
     "com.unity.collab-proxy": "2.7.1",
+    "com.unity.device-simulator.devices": "1.0.0",
     "com.unity.feature.mobile": "1.0.0",
     "com.unity.ide.rider": "3.0.31",
     "com.unity.ide.visualstudio": "2.0.22",
